prevently-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.4
firebase-admin==6.2.0
anthropic>=0.30.0